requirements.txt
streamlit
pandas
python-whois
openpyxl
xlrd
requests